Fix warnings

+++ src/utilities/actions/deleteTimelineEvent.ts
@@ -1,11 +1,16 @@
 'use server'
 
 import { ReportEvent } from '@/models/ReportEvent'
+import { auth } from '@clerk/nextjs'
 import { revalidatePath } from 'next/cache'
 import ReportEventCosmosClient from '../cosmosdb/ReportEventCosmosClient'
 
 export async function deleteTimelineEvent(eventId: string | undefined) {
   if (!eventId) return
+  const { userId } = auth()
+  if (!userId) {
+    return new Response('Unauthorized', { status: 401 })
+  }
 
   const eventCosmosClient = new ReportEventCosmosClient()
 
@@ -13,9 +18,13 @@ export async function deleteTimelineEvent(eventId: string | undefined) {
 
   if (!event || !Array.isArray(event)) return
 
+  if (userId !== event[0].userId) {
+    return new Response('Unauthorized', { status: 401 })
+  }
+
   await eventCosmosClient.deleteById(eventId, event[0].reportId)
 
-  revalidatePath(`/dashboard/reports/[reportSlug]`)
+  revalidatePath(`/dashboard/reports/[reportSlug]`, 'layout')
 
   return
 }
